requirements.txt
flask==2.0.2
gunicorn==20.1.0
boto3==1.22.0
mysql-connector-python==8.0.28
python-dotenv==0.20.0